requirements.txt
Django
Pillow

#code style and improvements
pre-commit
Unipath

#Secrets
python-decouple

# Storage
psycopg2
